PP-2939 Add get endpoint for /secure, similarly to dd frontend
 - Integrators might redirect to our service by calling the `next_url`, not just
   the `next_url_post`, so we are adding an endpoint to cater for this use case
